DRTII-1324 Move aggregated db class
